code: Wave 75-119 .cpp实现+代码质量修复—24文件
信道编码模块新Wave实现: LDPCCode5, TurboCode5, PolarCode8,
SpinalCode7, CascadeCode3, BchCode10, SpinalCode11, TurboCode9,
GrayCode9, CascadeCode7, LDPCCode7, ReedSolomon12, BchCode11,
PolarCode13, SpinalCode12, TurboCode10, GrayCode10, CascadeCode8
+ 现有LDPC/BCH/Polar/Erasure/Convolutional代码质量修复

评分: 427 + 1 = 428
变更: 24 files, 5753 insertions, 8 deletions

+++ src/utils/code47/ConvolutionalCode5.cpp
@@ -184,7 +184,7 @@ QVector<int> ConvolutionalCode5::decode(const QVector<double>& softBits)
     const int n = m_generators.size();
 
     /* 去打孔 */
-    QVector<double> depunctured = depuncture(softBits);
+    QVector<int> depunctured = depuncture(softBits);
     const int totalSymbols = depunctured.size();
     const int numSteps = totalSymbols / n;
